Minimalist join relation, needs client impl
As described in further detail in the comments, the join resource isn't as elegant as I'd like...
But it's better to err on the side of too separate until there's a better resource to combine it with.

I'm leaning towards fleshing out the concept of "current_round" and giving it participant relations.

The downside of this is that I really like how specific rounds are currently, for the most part, static... But maybe that's an oversimplification.