Add basic register.add api for player pool + specs
Still needs to be moved into html5.js

Seems like playerPosMap might be removable? Keeping it for now, it'll be a lot clearer after it's refactored

Main goal here is to remove all references to the renderer from html5.js... I think the main remaining targets are:

- Round change

- Player death

- Directing player death event

- Waiting events

So not too much stuff. Once the references are moved then a lot more cleanup can probably very easily happen as well